Make package so we can move the plethora of security bits out
It's better this way, much more understandable
