+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/SecurityConfig.java
@@ -16,97 +16,32 @@
  */
 package uk.ac.manchester.spinnaker.alloc;
 
-import static java.lang.String.format;
-import static java.time.Instant.now;
-import static java.util.Arrays.asList;
-import static java.util.Collections.unmodifiableCollection;
-import static java.util.Objects.isNull;
-import static java.util.Objects.nonNull;
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toSet;
-import static javax.net.ssl.TrustManagerFactory.getDefaultAlgorithm;
-import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
-import static javax.ws.rs.core.Response.status;
-import static javax.ws.rs.core.Response.Status.FORBIDDEN;
 import static org.slf4j.LoggerFactory.getLogger;
-import static org.springframework.http.HttpStatus.UNAUTHORIZED;
-import static uk.ac.manchester.spinnaker.alloc.AppAuthTransformationFilter.clearToken;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfig.TrustLevel.USER;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfigUtils.installInjectableTrustStoreAsDefault;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfigUtils.loadTrustStore;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfigUtils.trustManager;
+import static uk.ac.manchester.spinnaker.alloc.security.AppAuthTransformationFilter.clearToken;
+import static uk.ac.manchester.spinnaker.alloc.security.Utils.installInjectableTrustStoreAsDefault;
+import static uk.ac.manchester.spinnaker.alloc.security.Utils.loadTrustStore;
+import static uk.ac.manchester.spinnaker.alloc.security.Utils.trustManager;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.NotSerializableException;
-import java.io.ObjectOutputStream;
-import java.io.PrintWriter;
-import java.io.Serializable;
 import java.security.GeneralSecurityException;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.Principal;
-import java.security.SecureRandom;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.function.Supplier;
-import java.util.stream.Stream;
 
-import javax.annotation.PostConstruct;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.TrustManagerFactory;
 import javax.net.ssl.X509TrustManager;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
 
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.expression.EvaluationContext;
-import org.springframework.expression.Expression;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
-import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
-import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
@@ -115,11 +50,12 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.view.AbstractUrlBasedView;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
-import com.fasterxml.jackson.databind.json.JsonMapper;
-
 import uk.ac.manchester.spinnaker.alloc.ServiceConfig.URLPathMaker;
 import uk.ac.manchester.spinnaker.alloc.SpallocProperties.AuthProperties;
 import uk.ac.manchester.spinnaker.alloc.SpallocProperties.OpenIDProperties;
+import uk.ac.manchester.spinnaker.alloc.security.AppAuthTransformationFilter;
+import uk.ac.manchester.spinnaker.alloc.security.BasicAuthEntryPoint;
+import uk.ac.manchester.spinnaker.alloc.security.LocalAuthenticationProvider;
 
 /**
  * The security and administration configuration of the service.
@@ -154,27 +90,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 	 * in depth.
 	 */
 	public static final String IS_USER = "hasRole('USER')";
-
-	/**
-	 * The authority used to grant a user permission to create jobs, manipulate
-	 * them, and read their details. Note that many features are locked to the
-	 * owner of the job or admins. Users should also have {@link #GRANT_READER}.
-	 */
-	protected static final String GRANT_USER = "ROLE_USER";
-
-	/**
-	 * The authority used to grant a user permission to get general machine
-	 * information and summaries of jobs. Without this, only the service root
-	 * (and the parts required for logging in) will be visible.
-	 */
-	protected static final String GRANT_READER = "ROLE_READER";
-
-	/**
-	 * The authority used to grant a user permission to use administrative
-	 * actions. Admins should also have {@link #GRANT_USER} and
-	 * {@link #GRANT_READER}.
-	 */
-	protected static final String GRANT_ADMIN = "ROLE_ADMIN";
 
 	/** The name of the Spring MVC error view. */
 	public static final String MVC_ERROR = "erroroccurred";
@@ -247,87 +162,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 		auth.authenticationProvider(localAuthProvider);
 	}
 
-	/**
-	 * Just how trusted is a user?
-	 */
-	public enum TrustLevel {
-		/** Grants no real permissions at all. */
-		BASIC(),
-		/**
-		 * Grants read-only permissions in addition to {@link #BASIC}.
-		 *
-		 * @see SecurityConfig#GRANT_READER
-		 */
-		READER(GRANT_READER),
-		/**
-		 * Grants job creation and management permissions in addition to
-		 * {@link #READER}.
-		 *
-		 * @see SecurityConfig#GRANT_USER
-		 */
-		USER(GRANT_READER, GRANT_USER),
-		/**
-		 * Grants service administration permissions in addition to
-		 * {@link #USER}.
-		 *
-		 * @see SecurityConfig#GRANT_ADMIN
-		 */
-		ADMIN(GRANT_READER, GRANT_USER, GRANT_ADMIN);
-
-		private List<String> grants;
-
-		TrustLevel(String... grants) {
-			this.grants = asList(grants);
-		}
-
-		Stream<GrantedAuthority> getGrants() {
-			return grants.stream().map(SimpleGrantedAuthority::new);
-		}
-	}
-
-	/**
-	 * Locally-defined authentication providers include the capability to create
-	 * users.
-	 *
-	 * @author Donal Fellows
-	 */
-	public interface LocalAuthenticationProvider extends
-			org.springframework.security.authentication.AuthenticationProvider {
-		/**
-		 * Create a user. Only admins can create users.
-		 *
-		 * @param username
-		 *            The user name to use.
-		 * @param password
-		 *            The <em>unencoded</em> password to use.
-		 * @param trustLevel
-		 *            How much is the user trusted.
-		 * @param quota
-		 *            The user's quota, in board-seconds.
-		 * @return True if the user was created, false if the user already
-		 *         existed.
-		 */
-		@PreAuthorize(IS_ADMIN)
-		boolean createUser(String username, String password,
-				TrustLevel trustLevel, long quota);
-
-		/**
-		 * Unlock any locked users whose lock period has expired.
-		 */
-		void unlockLockedUsers();
-
-		/**
-		 * Convert the type of the authentication in the security context.
-		 *
-		 * @param ctx
-		 *            The security context.
-		 * @return The new authentication (which is also installed into the
-		 *         security context), or {@code null} <em>if the authentication
-		 *         is not changed</em>.
-		 */
-		Authentication updateAuthentication(SecurityContext ctx);
-	}
-
 	private String oidcPath(String suffix) {
 		return urlMaker.systemUrl("perform_oidc/" + suffix);
 	}
@@ -396,167 +230,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 				.logoutSuccessUrl(loginUrl);
 	}
 
-	/**
-	 * Implements basic and bearer auth challenge presentation.
-	 */
-	@Component
-	static class BasicAuthEntryPoint implements AuthenticationEntryPoint {
-		private static final String AUTH = "WWW-Authenticate";
-
-		@Autowired
-		private AuthProperties props;
-
-		private String basicChallenge() {
-			return format("Basic realm=\"%s\"", props.getRealm());
-		}
-
-		private String openidChallenge() {
-			Set<String> scopes = props.getOpenid().getScopes();
-			return format("Bearer realm=\"%s\", scope=\"%s\"", //
-					props.getRealm(), scopes.stream().collect(joining(", ")));
-		}
-
-		@Override
-		public void commence(HttpServletRequest request,
-				HttpServletResponse response, AuthenticationException authEx)
-				throws IOException {
-			log.info("issuing request for log in to {}",
-					request.getRemoteAddr());
-			/*
-			 * The two API auth methods should be separate headers; the specs
-			 * say they can be one, but that's an area where things get
-			 * "exciting" (i.e., ambiguous) so we don't!
-			 */
-			if (props.isBasic()) {
-				response.addHeader(AUTH, basicChallenge());
-			}
-			if (props.getOpenid().isEnable()) {
-				response.addHeader(AUTH, openidChallenge());
-			}
-			response.setStatus(SC_UNAUTHORIZED);
-
-			// Provide a basic body; NB, don't need to close the writer
-			response.setCharacterEncoding("UTF-8");
-			response.setContentType("text/plain");
-			PrintWriter writer = response.getWriter();
-			writer.println("log in required, "
-					+ "either with BASIC auth or HBP/EBRAINS bearer token "
-					+ "(if they are enabled)");
-			writer.flush(); // Commit the response
-		}
-	}
-
-	private static final String BLAND_AUTH_MSG = "computer says no";
-
-	/**
-	 * Contains a single basic role grant.
-	 */
-	static class SimpleGrantedAuthority implements GrantedAuthority {
-		private static final long serialVersionUID = -668405047103939708L;
-
-		private String role;
-
-		SimpleGrantedAuthority(String role) {
-			this.role = role;
-		}
-
-		@Override
-		public String getAuthority() {
-			return role;
-		}
-
-		@Override
-		public String toString() {
-			return role;
-		}
-	}
-
-	/**
-	 * Make access denied (from a {@code @}{@link PreAuthorize} check) not fill
-	 * the log with huge stack traces.
-	 */
-	@Component
-	@Provider
-	static class AccessDeniedExceptionMapper
-			implements ExceptionMapper<AccessDeniedException> {
-		@Context
-		private UriInfo ui;
-
-		@Context
-		private HttpServletRequest req;
-
-		@Override
-		public Response toResponse(AccessDeniedException exception) {
-			// Actually produce useful logging; the default is ghastly!
-			Principal p = req.getUserPrincipal();
-			if (p instanceof AbstractAuthenticationToken) {
-				AbstractAuthenticationToken who =
-						(AbstractAuthenticationToken) p;
-				log.warn("access denied: {} : {} {}", ui.getAbsolutePath(),
-						who.getName(),
-						who.getAuthorities().stream()
-								.map(GrantedAuthority::getAuthority)
-								.collect(toSet()));
-			} else if (p instanceof OAuth2AuthenticatedPrincipal) {
-				OAuth2AuthenticatedPrincipal who =
-						(OAuth2AuthenticatedPrincipal) p;
-				log.warn("access denied: {} : {} {}", ui.getAbsolutePath(),
-						who.getName(),
-						who.getAuthorities().stream()
-								.map(GrantedAuthority::getAuthority)
-								.collect(toSet()));
-			} else {
-				log.warn("access denied: {} : {}", ui.getAbsolutePath(), p);
-			}
-			// But the user gets a bland response
-			return status(FORBIDDEN).entity(BLAND_AUTH_MSG).build();
-		}
-	}
-
-	@Component
-	static class MyAuthenticationFailureHandler
-			implements AuthenticationFailureHandler {
-		@Autowired
-		private AuthProperties properties;
-
-		@Autowired
-		private JsonMapper mapper;
-
-		@Override
-		public void onAuthenticationFailure(HttpServletRequest request,
-				HttpServletResponse response, AuthenticationException e)
-				throws IOException, ServletException {
-			log.info("auth failure", e);
-			response.setStatus(UNAUTHORIZED.value());
-
-			String message = BLAND_AUTH_MSG;
-			if (properties.isDebugFailures()) {
-				message += ": " + e.getLocalizedMessage();
-			}
-			mapper.writeValue(response.getOutputStream(),
-					new AuthFailureObject(message));
-		}
-
-		static class AuthFailureObject {
-			private String message;
-
-			private Instant timestamp;
-
-			AuthFailureObject(String message) {
-				this.message = message;
-				this.timestamp = now();
-			}
-
-			public String getMessage() {
-				return message;
-			}
-
-			public Instant getTimestamp() {
-				return timestamp;
-			}
-		}
-	}
-
 	@Bean
 	LogoutHandler logoutHandler() {
 		SecurityContextLogoutHandler sclh = new SecurityContextLogoutHandler();
@@ -604,388 +277,5 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 					.addResourceLocations(
 							"classpath:/META-INF/public-web-resources/");
 		}
-	}
-
-
-	/**
-	 * Misc services related to password handling.
-	 *
-	 * @author Donal Fellows
-	 */
-	@Component
-	public static class PasswordServices {
-		private static final int BCRYPT_STRENGTH = 10;
-
-		private static final int PASSWORD_LENGTH = 16;
-
-		private PasswordEncoder passwordEncoder;
-
-		private SecureRandom rng;
-
-		@PostConstruct
-		private void init() {
-			rng = new SecureRandom();
-			passwordEncoder = new BCryptPasswordEncoder(BCRYPT_STRENGTH, rng);
-		}
-
-		/**
-		 * Generate a random password.
-		 *
-		 * @return A password consisting of 16 random ASCII printable
-		 *         characters.
-		 */
-		public final String generatePassword() {
-			StringBuilder sb = new StringBuilder();
-			rng.ints(PASSWORD_LENGTH, '\u0021', '\u007f')
-					.forEachOrdered(c -> sb.append((char) c));
-			return sb.toString();
-		}
-
-		/**
-		 * Encode a password with bcrypt. <em>This is a slow operation! Do not
-		 * hold a database transaction open when calling this.</em>
-		 *
-		 * @param password
-		 *            The password to encode. May be {@code null}.
-		 * @return The encoded password. Will be {@code null} if the input is
-		 *         {@code null}.
-		 */
-		public final String encodePassword(String password) {
-			return isNull(password) ? null : passwordEncoder.encode(password);
-		}
-
-		/**
-		 * Check using bcrypt if a password matches its encoded form retrieved
-		 * from the database. <em>This is a slow operation! Do not hold a
-		 * database transaction open when calling this.</em>
-		 *
-		 * @param password
-		 *            The password to check. If {@code null}, the encoded form
-		 *            must also be {@code null} for a match to be true.
-		 * @param encodedPassword
-		 *            The encoded form from the database.
-		 * @return True if the password matches, false otherwise.
-		 */
-		public final boolean matchPassword(String password,
-				String encodedPassword) {
-			if (isNull(password)) {
-				return isNull(encodedPassword);
-			}
-			return nonNull(encodedPassword)
-					&& passwordEncoder.matches(password, encodedPassword);
-		}
-	}
-
-	/**
-	 * Because Spring otherwise can't apply {@link PostFilter} to
-	 * {@link Stream}.
-	 *
-	 * @author Donal Fellows
-	 * @see <a href="https://stackoverflow.com/q/66107075/301832">Stack
-	 *      Overflow</a>
-	 */
-	@Component
-	static class AnyTypeMethodSecurityExpressionHandler
-			extends DefaultMethodSecurityExpressionHandler
-			implements MethodSecurityExpressionHandler {
-		@Override
-		public Object filter(Object target, Expression expr,
-				EvaluationContext ctx) {
-			if (isNull(target)) {
-				// We can handle this case here!
-				return null;
-			}
-			if (target instanceof Collection || target.getClass().isArray()
-					|| target instanceof Map || target instanceof Stream) {
-				return super.filter(target, expr, ctx);
-			}
-			if (target instanceof Optional) {
-				return filterOptional((Optional<?>) target, expr, ctx);
-			} else {
-				return filterOptional(Optional.of(target), expr, ctx)
-						.orElse(null);
-			}
-		}
-
-		@SuppressWarnings("unchecked")
-		private <T> Optional<T> filterOptional(Optional<T> target,
-				Expression expr, EvaluationContext ctx) {
-			// Java 8 language profile makes this a little messy
-			List<T> a = new ArrayList<T>();
-			target.ifPresent(a::add);
-			return ((Stream<T>) super.filter(a.stream(), expr, ctx))
-					.findFirst();
-		}
-	}
-
-	/**
-	 * Encodes what a user is permitted to do. Abstracts over several types of
-	 * security context.
-	 */
-	public static final class Permit {
-		/** Is the user an admin? */
-		public final boolean admin;
-
-		/** What is the name of the user? */
-		public final String name;
-
-		private List<GrantedAuthority> authorities = new ArrayList<>();
-
-		private static final String[] STDAUTH = {
-			GRANT_ADMIN, GRANT_READER, GRANT_USER
-		};
-
-		/**
-		 * Build a permit.
-		 *
-		 * @param context
-		 *            The originating security context.
-		 */
-		public Permit(javax.ws.rs.core.SecurityContext context) {
-			for (String role : STDAUTH) {
-				if (context.isUserInRole(role)) {
-					authorities.add(new SimpleGrantedAuthority(role));
-				}
-			}
-			admin = isAdmin(authorities);
-			name = context.getUserPrincipal().getName();
-		}
-
-		private static boolean isAdmin(List<GrantedAuthority> auths) {
-			return auths.stream().map(GrantedAuthority::getAuthority)
-					.anyMatch(GRANT_ADMIN::equals);
-		}
-
-		/**
-		 * Build a permit.
-		 *
-		 * @param context
-		 *            The originating security context.
-		 */
-		public Permit(SecurityContext context) {
-			authorities = new ArrayList<>(
-					context.getAuthentication().getAuthorities());
-			admin = isAdmin(authorities);
-			name = context.getAuthentication().getName();
-		}
-
-		/**
-		 * Build a permit for a service user. The service user can create jobs
-		 * and read job details, but cannot do much with jobs owned by other
-		 * users. <em>Only used by the legacy interface.</em>
-		 *
-		 * @param serviceUser
-		 *            The user name. Must exist in order to be actually used.
-		 */
-		public Permit(String serviceUser) {
-			USER.getGrants().forEach(authorities::add);
-			admin = false;
-			name = serviceUser;
-		}
-
-		/**
-		 * Can something owned by a given user can be shown to the user that
-		 * this permit is for?
-		 *
-		 * @param owner
-		 *            The owner of the object.
-		 * @return True exactly if the object (or subset of properties) may be
-		 *         shown.
-		 */
-		public boolean unveilFor(String owner) {
-			return admin || owner.equals(name);
-		}
-
-		/**
-		 * Push our special temporary authentication object for the duration of
-		 * the inner code. Used to satisfy Spring method security.
-		 *
-		 * @param <T>
-		 *            The type of the result
-		 * @param inContext
-		 *            The inner code to run with an authentication object
-		 *            applied.
-		 * @return Whatever the inner code returns
-		 */
-		public <T> T authorize(Supplier<T> inContext) {
-			/**
-			 * A temporarily-installable authentication token. Allows access to
-			 * secured APIs in asynchronous worker threads, provided they
-			 * provide a {@link Permit} (obtained from a service thread) to show
-			 * that they may do so.
-			 */
-			@SuppressWarnings("serial")
-			class TempAuth implements Authentication {
-				// The permit already proves we're authenticated
-				private boolean auth = true;
-
-				@Override
-				public String getName() {
-					return name;
-				}
-
-				@Override
-				public Collection<? extends GrantedAuthority> getAuthorities() {
-					return unmodifiableCollection(authorities);
-				}
-
-				@Override
-				public Object getCredentials() {
-					// You can never get the credentials from this
-					return null;
-				}
-
-				@Override
-				public Permit getDetails() {
-					return Permit.this;
-				}
-
-				@Override
-				public String getPrincipal() {
-					return name;
-				}
-
-				@Override
-				public boolean isAuthenticated() {
-					return auth;
-				}
-
-				@Override
-				public void setAuthenticated(boolean isAuthenticated) {
-					if (!isAuthenticated) {
-						auth = false;
-					}
-				}
-
-				private void writeObject(ObjectOutputStream out)
-						throws NotSerializableException {
-					throw new NotSerializableException(
-							"not actually serializable");
-				}
-			}
-
-			SecurityContext c = SecurityContextHolder.getContext();
-			Authentication old = c.getAuthentication();
-			c.setAuthentication(new TempAuth());
-			try {
-				return inContext.get();
-			} finally {
-				c.setAuthentication(old);
-			}
-		}
-	}
-}
-
-/** Support utility methods for working with SSL stuff. */
-abstract class SecurityConfigUtils {
-	private SecurityConfigUtils() {
-	}
-
-	/**
-	 * Load a trust store into a trust manager.
-	 *
-	 * @param truststore
-	 *            The trust store to load, or {@code null} to use the default.
-	 * @return The configured trust manager.
-	 * @throws GeneralSecurityException
-	 *             If things go wrong.
-	 * @see <a href="https://stackoverflow.com/a/24561444/301832">Stack
-	 *      Overflow</a>
-	 */
-	static X509TrustManager trustManager(KeyStore truststore)
-			throws GeneralSecurityException {
-		TrustManagerFactory tmf =
-				TrustManagerFactory.getInstance(getDefaultAlgorithm());
-		tmf.init(truststore);
-		for (TrustManager tm : tmf.getTrustManagers()) {
-			if (tm instanceof X509TrustManager) {
-				return (X509TrustManager) tm;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Set up the default SSL context so that we can inject our trust store for
-	 * verifying servers.
-	 *
-	 * @param injector
-	 *            How to get the value to inject.
-	 * @throws GeneralSecurityException
-	 *             If the SSL context can't be built.
-	 * @see <a href="https://stackoverflow.com/a/24561444/301832">Stack
-	 *      Overflow</a>
-	 */
-	static void installInjectableTrustStoreAsDefault(
-			Supplier<X509TrustManager> injector)
-			throws GeneralSecurityException {
-		X509TrustManager defaultTm = trustManager(null);
-		SSLContext sslContext = SSLContext.getInstance("TLS");
-		sslContext.init(null, new TrustManager[] {
-			new X509TrustManager() {
-				@Override
-				public X509Certificate[] getAcceptedIssuers() {
-					return defaultTm.getAcceptedIssuers();
-				}
-
-				@Override
-				public void checkServerTrusted(X509Certificate[] chain,
-						String authType) throws CertificateException {
-					X509TrustManager customTm = injector.get();
-					if (nonNull(customTm)) {
-						try {
-							customTm.checkServerTrusted(chain, authType);
-							// If we got here, we passed!
-							return;
-						} catch (CertificateException e) {
-						}
-					}
-					defaultTm.checkServerTrusted(chain, authType);
-				}
-
-				@Override
-				public void checkClientTrusted(X509Certificate[] chain,
-						String authType) throws CertificateException {
-					defaultTm.checkClientTrusted(chain, authType);
-				}
-			}
-		}, null);
-		SSLContext.setDefault(sslContext);
-	}
-
-	static KeyStore loadTrustStore(OpenIDProperties p) throws KeyStoreException,
-			IOException, NoSuchAlgorithmException, CertificateException {
-		KeyStore myTrustStore = KeyStore.getInstance(p.getTruststoreType());
-		try (InputStream myCerts = p.getTruststorePath().getInputStream()) {
-			myTrustStore.load(myCerts, p.getTruststorePassword().toCharArray());
-		}
-		return myTrustStore;
-	}
-}
-
-/**
- * A saved authentication token. Categorically only ever valid in the session
- * for which it was created; if the session changes, it cannot be reused.
- *
- * @author Donal Fellows
- */
-class Token implements Serializable {
-	private static final long serialVersionUID = -439034988839648948L;
-
-	private final String id;
-
-	private final Authentication auth;
-
-	Token(HttpSession s, Authentication a) {
-		this.auth = a;
-		this.id = s.getId();
-	}
-
-	boolean isValid(HttpSession s) {
-		return s.getId().equals(id);
-	}
-
-	Authentication getAuth() {
-		return auth;
 	}
 }
